requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
email_validator
Werkzeug
gunicorn
psycopg2-binary
Flask-Mail
python-dotenv
